+++ R/count.r
@@ -22,15 +22,10 @@
 #' str_count(c("a.", "...", ".a.a"), fixed("."))
 str_count <- function(string, pattern = "") {
   switch(type(pattern),
-    empty = stri_count_boundaries(string,
-      opts_brkiter = stri_opts_brkiter(type = "character")),
-    bound = stri_count_boundaries(string,
-      opts_brkiter = attr(pattern, "options")),
-    fixed = stri_count_fixed(string, pattern,
-      opts_fixed = attr(pattern, "options")),
-    coll  = stri_count_coll(string, pattern,
-      opts_collator = attr(pattern, "options")),
-    regex = stri_count_regex(string, pattern,
-      opts_regex = attr(pattern, "options"))
+    empty = stri_count_boundaries(string, opts_brkiter = opts(pattern)),
+    bound = stri_count_boundaries(string, opts_brkiter = opts(pattern)),
+    fixed = stri_count_fixed(string, pattern, opts_fixed = opts(pattern)),
+    coll  = stri_count_coll(string, pattern, opts_collator = opts(pattern)),
+    regex = stri_count_regex(string, pattern, opts_regex = opts(pattern))
   )
 }