+++ R/detect.r
@@ -35,12 +35,9 @@ str_detect <- function(string, pattern) {
   switch(type(pattern),
     empty = ,
     bound = str_count(string, pattern) > 0,
-    fixed = stri_detect_fixed(string, pattern,
-      opts_fixed = attr(pattern, "options")),
-    coll  = stri_detect_coll(string, pattern,
-      opts_collator = attr(pattern, "options")),
-    regex = stri_detect_regex(string, pattern,
-      opts_regex = attr(pattern, "options"))
+    fixed = stri_detect_fixed(string, pattern, opts_fixed = opts(pattern)),
+    coll  = stri_detect_coll(string,  pattern, opts_collator = opts(pattern)),
+    regex = stri_detect_regex(string, pattern, opts_regex = opts(pattern))
   )
 }