+++ R/extract.r
@@ -29,16 +29,11 @@
 #' str_extract_all("This is, suprisingly, a sentence.", boundary("word"))
 str_extract <- function(string, pattern) {
   switch(type(pattern),
-    empty = stri_extract_first_boundaries(string, pattern,
-      opts_brkiter = stri_opts_brkiter(type = "character")),
-    bound = stri_extract_first_boundaries(string, pattern,
-      opts_brkiter = attr(pattern, "options")),
-    fixed = stri_extract_first_fixed(string, pattern,
-      opts_fixed = attr(pattern, "options")),
-    coll  = stri_extract_first_coll(string, pattern,
-      opts_collator = attr(pattern, "options")),
-    regex = stri_extract_first_regex(string, pattern,
-      opts_regex = attr(pattern, "options"))
+    empty = stri_extract_first_boundaries(string, pattern, opts_brkiter = opts(pattern)),
+    bound = stri_extract_first_boundaries(string, pattern, opts_brkiter = opts(pattern)),
+    fixed = stri_extract_first_fixed(string, pattern, opts_fixed = opts(pattern)),
+    coll  = stri_extract_first_coll(string, pattern, opts_collator = opts(pattern)),
+    regex = stri_extract_first_regex(string, pattern, opts_regex = opts(pattern))
   )
 }
 
@@ -47,18 +42,14 @@ str_extract <- function(string, pattern) {
 str_extract_all <- function(string, pattern, simplify = FALSE) {
   switch(type(pattern),
     empty = stri_extract_all_boundaries(string, pattern,
-      simplify = simplify, omit_no_match = TRUE,
-      opts_brkiter = stri_opts_brkiter(type = "character")),
+      simplify = simplify, omit_no_match = TRUE, opts_brkiter = opts(pattern)),
     bound = stri_extract_all_boundaries(string, pattern,
-      simplify = simplify, omit_no_match = TRUE,
-      opts_brkiter = attr(pattern, "options")),
+      simplify = simplify, omit_no_match = TRUE, opts_brkiter = opts(pattern)),
     fixed = stri_extract_all_fixed(string, pattern,
-      opts_fixed = attr(pattern, "options")),
+      simplify = simplify, omit_no_match = TRUE, opts_fixed = opts(pattern)),
     coll  = stri_extract_all_coll(string, pattern,
-      simplify = simplify, omit_no_match = TRUE,
-      opts_collator = attr(pattern, "options")),
+      simplify = simplify, omit_no_match = TRUE, opts_collator = opts(pattern)),
     regex = stri_extract_all_regex(string, pattern,
-      simplify = simplify, omit_no_match = TRUE,
-      opts_regex = attr(pattern, "options"))
+      simplify = simplify, omit_no_match = TRUE, opts_regex = opts(pattern))
   )
 }