+++ R/locate.r
@@ -27,33 +27,25 @@
 #' str_locate_all(fruit, "")
 str_locate <- function(string, pattern) {
   switch(type(pattern),
-    empty = stri_locate_first_boundaries(string,
-      opts_brkiter = stri_opts_brkiter("character")),
-    bound = stri_locate_first_boundaries(string,
-      opts_brkiter = attr(pattern, "options")),
-    fixed = stri_locate_first_fixed(string, pattern,
-      opts_fixed = attr(pattern, "options")),
-    coll  = stri_locate_first_coll(string, pattern,
-      opts_collator = attr(pattern, "options")),
-    regex = stri_locate_first_regex(string, pattern,
-      opts_regex = attr(pattern, "options"))
+    empty = stri_locate_first_boundaries(string, opts_brkiter = opts(pattern)),
+    bound = stri_locate_first_boundaries(string, opts_brkiter = opts(pattern)),
+    fixed = stri_locate_first_fixed(string, pattern, opts_fixed = opts(pattern)),
+    coll  = stri_locate_first_coll(string, pattern, opts_collator = opts(pattern)),
+    regex = stri_locate_first_regex(string, pattern, opts_regex = opts(pattern))
   )
 }
 
 #' @rdname str_locate
 #' @export
 str_locate_all <- function(string, pattern) {
+  opts <- opts(pattern)
+
   switch(type(pattern),
-    empty = stri_locate_all_boundaries(string, omit_no_match = TRUE,
-      opts_brkiter = stri_opts_brkiter("character")),
-    bound = stri_locate_all_boundaries(string, omit_no_match = TRUE,
-      opts_brkiter = attr(pattern, "options")),
-    fixed = stri_locate_all_fixed(string, pattern, omit_no_match = TRUE,
-      opts_fixed = attr(pattern, "options")),
-    regex = stri_locate_all_regex(string, pattern,
-      omit_no_match = TRUE, opts_regex = attr(pattern, "options")),
-    coll  = stri_locate_all_coll(string, pattern,
-      omit_no_match = TRUE, opts_collator = attr(pattern, "options"))
+    empty = stri_locate_all_boundaries(string, omit_no_match = TRUE, opts_brkiter = opts),
+    bound = stri_locate_all_boundaries(string, omit_no_match = TRUE, opts_brkiter = opts),
+    fixed = stri_locate_all_fixed(string, pattern, omit_no_match = TRUE, opts_fixed = opts),
+    regex = stri_locate_all_regex(string, pattern, omit_no_match = TRUE, opts_regex = opts),
+    coll  = stri_locate_all_coll(string, pattern, omit_no_match = TRUE, opts_collator = opts)
   )
 }